+++ src/interfaces/media.ts
@@ -26,7 +26,7 @@ interface MediaTypeInfo {
 	convertedExtension?: string;
 }
 
-const mediaTypes: MediaTypeInfo[] = [
+const fileTypes: MediaTypeInfo[] = [
 
     { originalMime: "image/png", extension: "png", convertedMime: "image/webp", convertedExtension: "webp" },
     { originalMime: "image/jpg", extension: "jpg", convertedMime: "image/webp" , convertedExtension: "webp" },
@@ -40,31 +40,31 @@ const mediaTypes: MediaTypeInfo[] = [
     { originalMime: "video/mpeg", extension: "mpeg", convertedMime: "video/mp4" , convertedExtension: "mp4" },
     { originalMime: "video/webm", extension: "webm", convertedMime: "video/mp4" , convertedExtension: "mp4" },
 
-    { originalMime: "audio/mpeg", extension: "mp3", convertedMime: "audio/mpeg"},
-    { originalMime: "audio/mpg", extension: "mp3", convertedMime: "audio/mpeg" },
-    { originalMime: "audio/mpeg3", extension: "mp3", convertedMime: "audio/mpeg"},
-    { originalMime: "audio/mp3", extension: "mp3", convertedMime: "audio/mpeg" },
+    { originalMime: "audio/mpeg", extension: "mp3", convertedMime: "audio/mpeg",  convertedExtension: "mp3" },
+    { originalMime: "audio/mpg", extension: "mp3", convertedMime: "audio/mpeg" , convertedExtension: "mp3" },
+    { originalMime: "audio/mpeg3", extension: "mp3", convertedMime: "audio/mpeg", convertedExtension: "mp3" },
+    { originalMime: "audio/mp3", extension: "mp3", convertedMime: "audio/mpeg" , convertedExtension: "mp3" },
 
-    { originalMime: "application/pdf", extension: "pdf", convertedMime: "application/pdf" },
-	{ originalMime: "application/javascript", extension: "js", convertedMime: "application/javascript" },
-    { originalMime: "application/json", extension: "json", convertedMime: "application/json" },
-    { originalMime: "application/vnd.ms-fontobject", extension: "eot", convertedMime: "application/vnd.ms-fontobject" },
-    { originalMime: "application/yaml", extension: "yaml", convertedMime: "application/yaml" },
-	{ originalMime: "application/xml", extension: "xml", convertedMime: "application/xml" },
+    { originalMime: "application/pdf", extension: "pdf", convertedMime: "application/pdf", convertedExtension: "pdf" },
+	{ originalMime: "application/javascript", extension: "js", convertedMime: "application/javascript", convertedExtension: "js" },
+    { originalMime: "application/json", extension: "json", convertedMime: "application/json", convertedExtension: "json" },
+    { originalMime: "application/vnd.ms-fontobject", extension: "eot", convertedMime: "application/vnd.ms-fontobject" , convertedExtension: "eot" },
+    { originalMime: "application/yaml", extension: "yaml", convertedMime: "application/yaml" , convertedExtension: "yaml" },
+	{ originalMime: "application/xml", extension: "xml", convertedMime: "application/xml" , convertedExtension: "xml" },
 
-    { originalMime: "font/otf", extension: "otf", convertedMime: "font/otf" },
-    { originalMime: "font/ttf", extension: "ttf", convertedMime: "font/ttf" },
-    { originalMime: "font/woff", extension: "woff", convertedMime: "font/woff" },
-    { originalMime: "font/woff2", extension: "woff2", convertedMime: "font/woff2" },
+    { originalMime: "font/otf", extension: "otf", convertedMime: "font/otf" , convertedExtension: "otf" },
+    { originalMime: "font/ttf", extension: "ttf", convertedMime: "font/ttf" , convertedExtension: "ttf" },
+    { originalMime: "font/woff", extension: "woff", convertedMime: "font/woff" , convertedExtension: "woff" },
+    { originalMime: "font/woff2", extension: "woff2", convertedMime: "font/woff2" , convertedExtension: "woff2" },
 
-	{ originalMime: "text/html", extension: "map", convertedMime: "text/html" },
-    { originalMime: "text/markdown", extension: "md", convertedMime: "text/markdown" },
-    { originalMime: "text/css", extension: "css", convertedMime: "text/css" },
-    { originalMime: "text/x-handlebars-template", extension: "hbs", convertedMime: "text/x-handlebars-template" },
-    { originalMime: "text/plain", extension: "txt", convertedMime: "text/plain" },
-	{ originalMime: "text/yaml", extension: "yaml", convertedMime: "text/yaml" },
+	{ originalMime: "text/html", extension: "map", convertedMime: "text/html" , convertedExtension: "map" },
+    { originalMime: "text/markdown", extension: "md", convertedMime: "text/markdown" , convertedExtension: "md" },
+    { originalMime: "text/css", extension: "css", convertedMime: "text/css" , convertedExtension: "css" },
+    { originalMime: "text/x-handlebars-template", extension: "hbs", convertedMime: "text/x-handlebars-template" , convertedExtension: "hbs" },
+    { originalMime: "text/plain", extension: "txt", convertedMime: "text/plain" , convertedExtension: "txt" },
+	{ originalMime: "text/yaml", extension: "yaml", convertedMime: "text/yaml" , convertedExtension: "yaml" },
 
-	{ originalMime: "model/stl", extension: "stl", convertedMime: "model/stl" },
+	{ originalMime: "model/stl", extension: "stl", convertedMime: "model/stl" , convertedExtension: "stl" },
 ];
 
 interface FileData{
@@ -111,7 +111,7 @@ export {
 	FileData,
 	LegacyMediaReturnMessage,
 	MediaInfoReturnMessage,
-	mediaTypes,
+	fileTypes,
 	ResultMessage,
 	UploadTypes,
 	UploadStatus,